+++ lib/data/models/models.dart
@@ -9,4 +9,5 @@ export 'episode.dart';
 export 'epg_programme.dart';
 export 'watch_progress.dart';
 export 'favorite.dart';
+export 'media_detail.dart';
 export 'series_detail.dart';